src/requirements.txt
django-widget-tweaks==1.4.5
Django==2.2.8
pytz==2019.1
selenium==3.141.0
sqlparse==0.3.0
urllib3==1.25.3
glob3==0.0.1
gunicorn==19.9.0
psycopg2-binary==2.8.3
djangorestframework==3.11.0
markdown==3.1.1
django-filter==2.2.0